@@ -1,10 +1,9 @@
-'use strict';
-const WebSocketServer = require('ws').Server;
-const assert = require('power-assert');
-const ARIWebSocket = require('../../src/events/ARIWebSocket');
+"use strict";
+const WebSocketServer = require("ws").Server;
+const assert = require("power-assert");
+const ARIWebSocket = require("../../src/events/ARIWebSocket");
 
-describe('ARIWebSocket', () => {
-
+describe("ARIWebSocket", () => {
   let ws;
   let server;
 
@@ -20,23 +19,23 @@ describe('ARIWebSocket', () => {
   });
 
   it('passes non-"message" events along without parsing', done => {
-    ws = new ARIWebSocket({ url: 'ws://localhost:8088' });
+    ws = new ARIWebSocket({ url: "ws://localhost:8088" });
 
-    ws.once('open', () => {
-      ws.once('bugaboo', message => {
-        assert.equal(message, 'bugaloo');
+    ws.once("open", () => {
+      ws.once("bugaboo", message => {
+        assert.equal(message, "bugaloo");
         done();
       });
 
-      ws.emit('bugaboo', 'bugaloo');
+      ws.emit("bugaboo", "bugaloo");
     });
   });
 
   it('passes "message" events without a payload along without parsing', done => {
-    ws = new ARIWebSocket({ url: 'ws://localhost:8088' });
+    ws = new ARIWebSocket({ url: "ws://localhost:8088" });
 
-    server.once('connection', conn => {
-      ws.on('message', message => {
+    server.once("connection", conn => {
+      ws.on("message", message => {
         assert.deepEqual(message, new Uint8Array(0));
         done();
       });
@@ -46,12 +45,12 @@ describe('ARIWebSocket', () => {
   });
 
   it('passes "message" events that are not string along without parsing', done => {
-    ws = new ARIWebSocket({ url: 'ws://localhost:8088' });
-    const buf = new Buffer('string');
+    ws = new ARIWebSocket({ url: "ws://localhost:8088" });
+    const buf = Buffer.from("string");
 
-    server.once('connection', conn => {
-      ws.once('message', message => {
-        assert.equal(message.toString('utf-8'), buf.toString('utf-8'));
+    server.once("connection", conn => {
+      ws.once("message", message => {
+        assert.equal(message.toString("utf-8"), buf.toString("utf-8"));
         done();
       });
 
@@ -60,29 +59,28 @@ describe('ARIWebSocket', () => {
   });
 
   it('passes string "message" events that cannot be parsed as JSON along as-is', done => {
-    ws = new ARIWebSocket({ url: 'ws://localhost:8088' });
+    ws = new ARIWebSocket({ url: "ws://localhost:8088" });
 
-    server.once('connection', conn => {
-      ws.once('message', message => {
-        assert.equal(message, 'hidey ho');
+    server.once("connection", conn => {
+      ws.once("message", message => {
+        assert.equal(message, "hidey ho");
         done();
       });
 
-      conn.send('hidey ho');
+      conn.send("hidey ho");
     });
   });
 
   it('parses JSON "message" payloads before triggering the event', done => {
-    ws = new ARIWebSocket({ url: 'ws://localhost:8088' });
+    ws = new ARIWebSocket({ url: "ws://localhost:8088" });
 
-    server.once('connection', conn => {
-      ws.once('message', message => {
-        assert.deepEqual(message, { foo: 'bar' });
+    server.once("connection", conn => {
+      ws.once("message", message => {
+        assert.deepEqual(message, { foo: "bar" });
         done();
       });
 
       conn.send('{"foo": "bar"}');
     });
   });
-
 });
