@@ -1,5 +1,5 @@
-'use strict';
-const Events = require('./events');
-const API = require('./api');
+"use strict";
+const Events = require("./events");
+const API = require("./api");
 
 module.exports = { Events, API };
