@@ -1,9 +1,8 @@
-'use strict';
-const EventsAPI = require('../../src/api/EventsAPI');
-const nock = require('nock');
+"use strict";
+const EventsAPI = require("../../src/api/EventsAPI");
+const nock = require("nock");
 
-describe('the Events API', () => {
-
+describe("the Events API", () => {
   beforeEach(() => {
     nock.disableNetConnect();
   });
@@ -12,33 +11,34 @@ describe('the Events API', () => {
     nock.enableNetConnect();
   });
 
-  describe('generateUserEvent method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .post('/ari/events/user/fooCreated', {
-          variables: { var1: 'cool', var2: 'excellent' }
+  describe("generateUserEvent method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .post("/ari/events/user/fooCreated", {
+          variables: { var1: "cool", var2: "excellent" }
         })
         .query({
-          application: 'fooApp',
-          source: 'endpoint:pjsip/6001,deviceStatus:someDevice'
+          application: "fooApp",
+          source: "endpoint:pjsip/6001,deviceStatus:someDevice"
         })
-        .reply(200, { foo: 'bar' });
+        .reply(200, { foo: "bar" });
 
       const api = new EventsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.generateUserEvent({
-        eventName: 'fooCreated',
-        application: 'fooApp',
-        source: ['endpoint:pjsip/6001', 'deviceStatus:someDevice'],
-        variables: { var1: 'cool', var2: 'excellent' }
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .generateUserEvent({
+          eventName: "fooCreated",
+          application: "fooApp",
+          source: ["endpoint:pjsip/6001", "deviceStatus:someDevice"],
+          variables: { var1: "cool", var2: "excellent" }
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 });
