@@ -1,9 +1,8 @@
-'use strict';
-const MailboxesAPI = require('../../src/api/MailboxesAPI');
-const nock = require('nock');
+"use strict";
+const MailboxesAPI = require("../../src/api/MailboxesAPI");
+const nock = require("nock");
 
-describe('the Mailboxes API', () => {
-
+describe("the Mailboxes API", () => {
   beforeEach(() => {
     nock.disableNetConnect();
   });
@@ -12,17 +11,16 @@ describe('the Mailboxes API', () => {
     nock.enableNetConnect();
   });
 
-  describe('list method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/mailboxes')
-        .reply(200, { foo: 'bar' });
+  describe("list method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/mailboxes")
+        .reply(200, { foo: "bar" });
 
       const api = new MailboxesAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
       return api.list().then(() => {
@@ -31,69 +29,72 @@ describe('the Mailboxes API', () => {
     });
   });
 
-  describe('get method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/mailboxes/foo')
-        .reply(200, { foo: 'bar' });
+  describe("get method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/mailboxes/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new MailboxesAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.get({
-        mailboxName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .get({
+          mailboxName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('update method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .put('/ari/mailboxes/foo')
+  describe("update method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .put("/ari/mailboxes/foo")
         .query({ oldMessages: 5, newMessages: 0 })
-        .reply(200, { foo: 'bar' });
+        .reply(200, { foo: "bar" });
 
       const api = new MailboxesAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.update({
-        mailboxName: 'foo',
-        oldMessages: 5,
-        newMessages: 0
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .update({
+          mailboxName: "foo",
+          oldMessages: 5,
+          newMessages: 0
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('destroy method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .delete('/ari/mailboxes/foo')
-        .reply(200, { foo: 'bar' });
+  describe("destroy method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .delete("/ari/mailboxes/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new MailboxesAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.destroy({
-        mailboxName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .destroy({
+          mailboxName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 });
