@@ -1,9 +1,8 @@
-'use strict';
-const RecordingsAPI = require('../../src/api/RecordingsAPI');
-const nock = require('nock');
+"use strict";
+const RecordingsAPI = require("../../src/api/RecordingsAPI");
+const nock = require("nock");
 
-describe('the Recordings API', () => {
-
+describe("the Recordings API", () => {
   beforeEach(() => {
     nock.disableNetConnect();
   });
@@ -12,17 +11,16 @@ describe('the Recordings API', () => {
     nock.enableNetConnect();
   });
 
-  describe('listStored method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/recordings/stored')
-        .reply(200, { foo: 'bar' });
+  describe("listStored method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/recordings/stored")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
       return api.listStored().then(() => {
@@ -31,236 +29,247 @@ describe('the Recordings API', () => {
     });
   });
 
-  describe('getStored method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/recordings/stored/foo')
-        .reply(200, { foo: 'bar' });
+  describe("getStored method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/recordings/stored/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.getStored({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .getStored({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('destroyStored method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .delete('/ari/recordings/stored/foo')
-        .reply(200, { foo: 'bar' });
+  describe("destroyStored method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .delete("/ari/recordings/stored/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.destroyStored({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .destroyStored({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('getStoredFile method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/recordings/stored/foo/file')
-        .reply(200, { foo: 'bar' });
+  describe("getStoredFile method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/recordings/stored/foo/file")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.getStoredFile({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .getStoredFile({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('copyStored method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .post('/ari/recordings/stored/foo/copy')
-        .query({ destinationRecordingName: 'bar' })
-        .reply(200, { foo: 'bar' });
+  describe("copyStored method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .post("/ari/recordings/stored/foo/copy")
+        .query({ destinationRecordingName: "bar" })
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.copyStored({
-        recordingName: 'foo',
-        destinationRecordingName: 'bar'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .copyStored({
+          recordingName: "foo",
+          destinationRecordingName: "bar"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('getLive method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/recordings/live/foo')
-        .reply(200, { foo: 'bar' });
+  describe("getLive method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/recordings/live/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.getLive({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .getLive({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('cancel method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .delete('/ari/recordings/live/foo')
-        .reply(200, { foo: 'bar' });
+  describe("cancel method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .delete("/ari/recordings/live/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.cancel({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .cancel({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('stop method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .post('/ari/recordings/live/foo/stop')
-        .reply(200, { foo: 'bar' });
+  describe("stop method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .post("/ari/recordings/live/foo/stop")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.stop({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .stop({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('pause method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .post('/ari/recordings/live/foo/pause')
-        .reply(200, { foo: 'bar' });
+  describe("pause method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .post("/ari/recordings/live/foo/pause")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.pause({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .pause({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('unpause method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .delete('/ari/recordings/live/foo/pause')
-        .reply(200, { foo: 'bar' });
+  describe("unpause method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .delete("/ari/recordings/live/foo/pause")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.unpause({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .unpause({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('mute method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .post('/ari/recordings/live/foo/mute')
-        .reply(200, { foo: 'bar' });
+  describe("mute method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .post("/ari/recordings/live/foo/mute")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.mute({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .mute({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('unmute method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .delete('/ari/recordings/live/foo/mute')
-        .reply(200, { foo: 'bar' });
+  describe("unmute method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .delete("/ari/recordings/live/foo/mute")
+        .reply(200, { foo: "bar" });
 
       const api = new RecordingsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.unmute({
-        recordingName: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .unmute({
+          recordingName: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 });
