@@ -1,9 +1,8 @@
-'use strict';
-const SoundsAPI = require('../../src/api/SoundsAPI');
-const nock = require('nock');
+"use strict";
+const SoundsAPI = require("../../src/api/SoundsAPI");
+const nock = require("nock");
 
-describe('the Sounds API', () => {
-
+describe("the Sounds API", () => {
   beforeEach(() => {
     nock.disableNetConnect();
   });
@@ -12,47 +11,49 @@ describe('the Sounds API', () => {
     nock.enableNetConnect();
   });
 
-  describe('list method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/sounds')
-        .query({ lang: 'en', format: 'ogg' })
-        .reply(200, { foo: 'bar' });
+  describe("list method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/sounds")
+        .query({ lang: "en", format: "ogg" })
+        .reply(200, { foo: "bar" });
 
       const api = new SoundsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.list({
-        lang: 'en',
-        format: 'ogg'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .list({
+          lang: "en",
+          format: "ogg"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 
-  describe('get method', () => {
-
-    it('makes the right request', () => {
-      const mock = nock('http://fake.local')
-        .get('/ari/sounds/foo')
-        .reply(200, { foo: 'bar' });
+  describe("get method", () => {
+    it("makes the right request", () => {
+      const mock = nock("http://fake.local")
+        .get("/ari/sounds/foo")
+        .reply(200, { foo: "bar" });
 
       const api = new SoundsAPI({
-        baseUrl: 'http://fake.local/ari',
-        username: 'user',
-        password: '1234'
+        baseUrl: "http://fake.local/ari",
+        username: "user",
+        password: "1234"
       });
 
-      return api.get({
-        soundId: 'foo'
-      }).then(() => {
-        mock.done();
-      });
+      return api
+        .get({
+          soundId: "foo"
+        })
+        .then(() => {
+          mock.done();
+        });
     });
   });
 });
