@@ -1,13 +1,10 @@
-'use strict';
-const WebSocketServer = require('ws').Server;
-const assert = require('power-assert');
-const ReconnectingWebSocket =
-  require('../../src/events/ReconnectingWebSocket');
+"use strict";
+const WebSocketServer = require("ws").Server;
+const assert = require("power-assert");
+const ReconnectingWebSocket = require("../../src/events/ReconnectingWebSocket");
 
-describe('ReconnectingWebSocket', () => {
-
-  describe('by default', () => {
-
+describe("ReconnectingWebSocket", () => {
+  describe("by default", () => {
     let ws;
     let server;
 
@@ -22,10 +19,10 @@ describe('ReconnectingWebSocket', () => {
       server.close();
     });
 
-    it('connects upon creation', done => {
-      ws = new ReconnectingWebSocket({ url: 'ws://localhost:8088' });
-      server.on('connection', conn => {
-        ws.on('open', () => {
+    it("connects upon creation", done => {
+      ws = new ReconnectingWebSocket({ url: "ws://localhost:8088" });
+      server.on("connection", conn => {
+        ws.on("open", () => {
           assert(conn);
           done();
         });
@@ -33,48 +30,49 @@ describe('ReconnectingWebSocket', () => {
     });
 
     it('emits an "open" event on connection', done => {
-      ws = new ReconnectingWebSocket({ url: 'ws://localhost:8088' });
-      ws.once('open', done);
+      ws = new ReconnectingWebSocket({ url: "ws://localhost:8088" });
+      ws.once("open", done);
     });
 
-    it('disconnects when close is called', done => {
+    it("disconnects when close is called", done => {
       ws = new ReconnectingWebSocket({
-        url: 'ws://localhost:8088',
+        url: "ws://localhost:8088",
         reconnect: false
       });
 
-      server.once('connection', conn => {
-        conn.once('close', () => {
+      server.once("connection", conn => {
+        conn.once("close", () => {
           done();
         });
       });
 
-      ws.once('open', () => ws.close());
+      ws.once("open", () => ws.close());
     });
 
     it('emits a "message" event when a message is received', done => {
-      ws = new ReconnectingWebSocket({ url: 'ws://localhost:8088' });
+      ws = new ReconnectingWebSocket({ url: "ws://localhost:8088" });
 
-      server.once('connection', conn => {
-        ws.once('message', message => {
-          assert.equal(message, 'hidey ho');
+      server.once("connection", conn => {
+        ws.once("message", message => {
+          assert.equal(message, "hidey ho");
           done();
         });
 
-        conn.send('hidey ho');
+        conn.send("hidey ho");
       });
     });
   });
 
-  describe('when no connection can be established on create', () => {
-
+  describe("when no connection can be established on create", () => {
     let ws;
     let server;
 
     beforeEach(() => {
       server = new WebSocketServer({
         port: 8088,
-        verifyClient: (info, cb) => { cb(false, 666, 'fail'); }
+        verifyClient: (info, verify) => {
+          verify(false, 666, "fail");
+        }
       });
     });
 
@@ -87,11 +85,11 @@ describe('ReconnectingWebSocket', () => {
 
     it('emits "error" event', done => {
       ws = new ReconnectingWebSocket({
-        url: 'ws://localhost:8088',
+        url: "ws://localhost:8088",
         reconnect: false
       });
 
-      ws.on('error', err => {
+      ws.on("error", err => {
         assert(err);
         done();
       });
@@ -99,19 +97,17 @@ describe('ReconnectingWebSocket', () => {
 
     it('emits "close" event', done => {
       ws = new ReconnectingWebSocket({
-        url: 'ws://localhost:8088',
+        url: "ws://localhost:8088",
         reconnect: false
       });
 
-      ws.on('error', () => {});
-      ws.on('close', done);
+      ws.on("error", () => {});
+      ws.on("close", done);
     });
   });
 
-  describe('when an error occurs after being connected', () => {
-
-    describe('and reconnect is disabled', () => {
-
+  describe("when an error occurs after being connected", () => {
+    describe("and reconnect is disabled", () => {
       let ws;
       let server;
 
@@ -128,16 +124,16 @@ describe('ReconnectingWebSocket', () => {
 
       it('emits an "error" event', done => {
         ws = new ReconnectingWebSocket({
-          url: 'ws://localhost:8088',
+          url: "ws://localhost:8088",
           reconnect: false
         });
 
-        ws.on('open', () => {
+        ws.on("open", () => {
           // TODO: come up with a better way to emulate this
-          ws.ws.emit('error', new Error('dangit'));
+          ws.ws.emit("error", new Error("dangit"));
         });
 
-        ws.once('error', err => {
+        ws.once("error", err => {
           assert(err);
           done();
         });
@@ -145,24 +141,22 @@ describe('ReconnectingWebSocket', () => {
 
       it('emits a "close" event', done => {
         ws = new ReconnectingWebSocket({
-          url: 'ws://localhost:8088',
+          url: "ws://localhost:8088",
           reconnect: false
         });
 
-        ws.on('open', () => {
+        ws.on("open", () => {
           // TODO: come up with a better way to emulate this
-          ws.ws.emit('error', new Error('dangit'));
+          ws.ws.emit("error", new Error("dangit"));
         });
 
-        ws.on('error', () => {});
-        ws.once('close', done);
+        ws.on("error", () => {});
+        ws.once("close", done);
       });
     });
 
-    describe('and reconnect is enabled', () => {
-
-      describe('and the server allows the reconnect', () => {
-
+    describe("and reconnect is enabled", () => {
+      describe("and the server allows the reconnect", () => {
         let ws;
         let server;
 
@@ -177,33 +171,31 @@ describe('ReconnectingWebSocket', () => {
           server.close();
         });
 
-        it('reconnects', done => {
+        it("reconnects", done => {
           let connectCount = 0;
           let reconnected = false;
 
-          ws = new ReconnectingWebSocket({ url: 'ws://localhost:8088' });
+          ws = new ReconnectingWebSocket({ url: "ws://localhost:8088" });
 
-          server.on('connection', () => {
+          server.on("connection", () => {
             connectCount++;
-            reconnected = (connectCount === 2);
+            reconnected = connectCount === 2;
           });
 
-          ws.on('open', () => {
+          ws.on("open", () => {
             // TODO: come up with a better way to emulate this
-            ws.ws.emit('error', new Error('dangit'));
+            ws.ws.emit("error", new Error("dangit"));
           });
 
-          ws.on('reconnected', () => {
+          ws.on("reconnected", () => {
             assert(reconnected);
             done();
           });
         });
       });
 
-      describe('but the reconnect fails', () => {
-
-        describe('and retries are disabled', () => {
-
+      describe("but the reconnect fails", () => {
+        describe("and retries are disabled", () => {
           let ws;
           let server;
 
@@ -211,9 +203,9 @@ describe('ReconnectingWebSocket', () => {
             let numConnectionAttempts = 0;
             server = new WebSocketServer({
               port: 8088,
-              verifyClient: (info, callback) => {
+              verifyClient: (info, verify) => {
                 numConnectionAttempts++;
-                callback(numConnectionAttempts === 1, 666, 'fail');
+                verify(numConnectionAttempts === 1, 666, "fail");
               }
             });
           });
@@ -227,16 +219,16 @@ describe('ReconnectingWebSocket', () => {
 
           it('emits an "error" event', done => {
             ws = new ReconnectingWebSocket({
-              url: 'ws://localhost:8088',
+              url: "ws://localhost:8088",
               retryOptions: { retries: 0 }
             });
 
-            ws.on('open', () => {
+            ws.on("open", () => {
               // TODO: come up with a better way to emulate this
-              ws.ws.emit('error', new Error('dangit'));
+              ws.ws.emit("error", new Error("dangit"));
             });
 
-            ws.once('error', err => {
+            ws.once("error", err => {
               assert(err);
               done();
             });
@@ -244,24 +236,23 @@ describe('ReconnectingWebSocket', () => {
 
           it('emits a "close" event', done => {
             ws = new ReconnectingWebSocket({
-              url: 'ws://localhost:8088',
+              url: "ws://localhost:8088",
               retryOptions: { retries: 0 }
             });
 
-            ws.on('open', () => {
+            ws.on("open", () => {
               // TODO: come up with a better way to emulate this
-              ws.ws.emit('error', new Error('dangit'));
+              ws.ws.emit("error", new Error("dangit"));
             });
 
-            ws.on('error', () => {});
-            ws.once('close', () => {
+            ws.on("error", () => {});
+            ws.once("close", () => {
               done();
             });
           });
         });
 
-        describe('but retries are enabled and the retry succeeds', () => {
-
+        describe("but retries are enabled and the retry succeeds", () => {
           let ws;
           let server;
 
@@ -269,9 +260,9 @@ describe('ReconnectingWebSocket', () => {
             let numConnectionAttempts = 0;
             server = new WebSocketServer({
               port: 8088,
-              verifyClient: (info, callback) => {
+              verifyClient: (info, verify) => {
                 numConnectionAttempts++;
-                callback(numConnectionAttempts % 2 === 0, 666, 'fail');
+                verify(numConnectionAttempts % 2 === 0, 666, "fail");
               }
             });
           });
@@ -283,18 +274,18 @@ describe('ReconnectingWebSocket', () => {
             server.close();
           });
 
-          it('reconnects', done => {
+          it("reconnects", done => {
             ws = new ReconnectingWebSocket({
-              url: 'ws://localhost:8088',
+              url: "ws://localhost:8088",
               retryOptions: { retries: 1 }
             });
 
-            ws.on('open', () => {
+            ws.on("open", () => {
               // TODO: come up with a better way to emulate this
-              ws.ws.emit('error', new Error('dangit'));
+              ws.ws.emit("error", new Error("dangit"));
             });
 
-            ws.once('reconnected', () => {
+            ws.once("reconnected", () => {
               done();
             });
           });
@@ -303,10 +294,8 @@ describe('ReconnectingWebSocket', () => {
     });
   });
 
-  describe('when the server closes the connection', () => {
-
-    describe('and reconnect is disabled', () => {
-
+  describe("when the server closes the connection", () => {
+    describe("and reconnect is disabled", () => {
       let ws;
       let server;
 
@@ -323,24 +312,22 @@ describe('ReconnectingWebSocket', () => {
 
       it('emits a "close" event', done => {
         ws = new ReconnectingWebSocket({
-          url: 'ws://localhost:8088',
+          url: "ws://localhost:8088",
           reconnect: false
         });
 
-        server.once('connection', conn => {
-          ws.once('open', () => {
+        server.once("connection", conn => {
+          ws.once("open", () => {
             conn.close();
           });
         });
 
-        ws.once('close', done);
+        ws.once("close", done);
       });
     });
 
-    describe('and reconnect is enabled', () => {
-
-      describe('and the server allows the reconnect', () => {
-
+    describe("and reconnect is enabled", () => {
+      describe("and the server allows the reconnect", () => {
         let ws;
         let server;
 
@@ -355,31 +342,30 @@ describe('ReconnectingWebSocket', () => {
           server.close();
         });
 
-        it('reconnects', done => {
+        it("reconnects", done => {
           let connectCount = 0;
           let reconnected = false;
 
-          ws = new ReconnectingWebSocket({ url: 'ws://localhost:8088' });
-          server.on('connection', conn => {
+          ws = new ReconnectingWebSocket({ url: "ws://localhost:8088" });
+          server.on("connection", conn => {
             connectCount++;
-            reconnected = (connectCount === 2);
+            reconnected = connectCount === 2;
 
             if (connectCount === 1) {
-              ws.once('open', () => {
+              ws.once("open", () => {
                 conn.close();
               });
             }
           });
 
-          ws.once('reconnected', () => {
+          ws.once("reconnected", () => {
             assert(reconnected);
             done();
           });
         });
       });
 
-      describe('but the reconnect fails', () => {
-
+      describe("but the reconnect fails", () => {
         let ws;
         let server;
 
@@ -387,9 +373,9 @@ describe('ReconnectingWebSocket', () => {
           let numConnectionAttempts = 0;
           server = new WebSocketServer({
             port: 8088,
-            verifyClient: (info, callback) => {
+            verifyClient: (info, verify) => {
               numConnectionAttempts++;
-              callback(numConnectionAttempts === 1, 666, 'fail');
+              verify(numConnectionAttempts === 1, 666, "fail");
             }
           });
         });
@@ -403,17 +389,17 @@ describe('ReconnectingWebSocket', () => {
 
         it('emits an "error" event', done => {
           ws = new ReconnectingWebSocket({
-            url: 'ws://localhost:8088',
+            url: "ws://localhost:8088",
             retryOptions: { retries: 0 }
           });
 
-          server.once('connection', conn => {
-            ws.once('open', () => {
+          server.once("connection", conn => {
+            ws.once("open", () => {
               conn.close();
             });
           });
 
-          ws.once('error', err => {
+          ws.once("error", err => {
             assert(err);
             done();
           });
@@ -421,18 +407,18 @@ describe('ReconnectingWebSocket', () => {
 
         it('emits a "close" event', done => {
           ws = new ReconnectingWebSocket({
-            url: 'ws://localhost:8088',
+            url: "ws://localhost:8088",
             retryOptions: { retries: 0 }
           });
 
-          server.once('connection', conn => {
-            ws.once('open', () => {
+          server.once("connection", conn => {
+            ws.once("open", () => {
               conn.close();
             });
           });
 
-          ws.on('error', () => {});
-          ws.once('close', () => {
+          ws.on("error", () => {});
+          ws.once("close", () => {
             done();
           });
         });
